rename `ZIP` to `QIT_CLI\Woo\ZipValidator`. Move to namespace `Woo` and make it a service
