+++ src/src/Commands/WooValidateZipCommand.php
@@ -2,14 +2,22 @@
 
 namespace QIT_CLI\Commands;
 
-use QIT_CLI\Zip;
+use QIT_CLI\Woo\ZipValidator;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Input\InputArgument;
 
 class WooValidateZipCommand extends Command {
 	protected static $defaultName = 'woo:validate-zip'; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.PropertyNotSnakeCase
+
+	private $zip_validator;
+
+	public function __construct( ZipValidator $zip_validator ) {
+		$this->zip_validator = $zip_validator;
+
+		parent::__construct();
+	}
 
 	protected function configure() {
 		$this
@@ -22,11 +30,12 @@ class WooValidateZipCommand extends Command {
 		try {
 			$zip_file = $input->getArgument( 'path' );
 
-			ZIP::validate_zip( $zip_file );
+			$this->zip_validator->validate_zip( $zip_file );
 
 			$output->writeln( '<info>Zip file content is valid.</info>' );
 		} catch ( \Exception $e ) {
-			$output->writeln( sprintf( '<error>An error occurred while validating the ZIP file. Error: %s</error>', $e->getMessage() ) );
+			$output->writeln( sprintf( '<error>An error occurred while validating the ZIP file. Error: %s</error>',
+			$e->getMessage() ) );
 
 			return Command::FAILURE;
 		}
